refactored the profile routing
